% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictml.R
\docType{package}
\name{ictml-package}
\alias{ictml}
\alias{ictml-package}
\title{ictml: Easily Install and Load the 'ICTML'}
\description{
The 'ICTML' is a software suite implementing the Interpretable Causal Targeted Machine Learning ('ICTML') Project \url{https://www.ictml.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.ictml.org}
}

}
\author{
\strong{Maintainer}: Henrik Bengtsson \email{henrikb@braju.com} [copyright holder]

}
\keyword{internal}
