% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictml_deps.R
\name{ictml_deps}
\alias{ictml_deps}
\title{List all ICTML dependencies}
\usage{
ictml_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidyverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\value{
A \link[base:data.frame]{base::data.frame} with columns \code{package}, \code{cran}, \code{local}, and \code{behind}.
}
\description{
List all ICTML dependencies
}
