% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictml_packages.R
\name{ictml_packages}
\alias{ictml_packages}
\title{List all packages in the ICTML}
\usage{
ictml_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include \strong{ictml} in the list?}
}
\value{
Returns a character vector of package names.
}
\description{
List all packages in the ICTML
}
\examples{
ictml_packages()

}
