% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictml_update.R
\name{ictml_update}
\alias{ictml_update}
\title{Update ICTML packages}
\usage{
ictml_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidyverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\value{
Returns nothing; outputs installation instructions, if needed.
}
\description{
This will check to see if all ICTML packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
ictml_update()
}

}
